# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TestUsil2021",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://testusil2021.tk/api",
            "description": "Test Usil",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        },
        {
            "url": "http://comicomerce.com/api",
            "description": "comicomerce.com",
            "variables": {
                "schema": {
                    "enum": [
                        "http"
                    ],
                    "default": "http"
                }
            }
        },
        {
            "url": "http://127.0.0.1/api",
            "description": "localhost",
            "variables": {
                "schema": {
                    "enum": [
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
        "/certificates": {
            "get": {
                "tags": [
                    "Certificates"
                ],
                "summary": "return certificates list",
                "responses": {
                    "200": {
                        "description": "Certificates List"
                    },
                    "default": {
                        "description": "Something went wrong"
                    }
                },
                "deprecated": false
            },
            "post": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Store a Certificate",
                "operationId": "cf74c9ca439d55af7fa9f94b61b9b88f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "certificate_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "course_id": "1",
                                    "name": "Jhon Doe",
                                    "certificate_date": "2021-11-29"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        },
        "/certificates/{id}": {
            "put": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Modifiy Certificates",
                "operationId": "fb6c06038853ba8cf7de563012510d6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Certificate ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "course_id": {
                                        "type": "string"
                                    },
                                    "certificate_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "course_id": "1",
                                    "name": "Jhon Doe",
                                    "certificate_date": "2021-11-29"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": [
                    "Certificates"
                ],
                "summary": "Delete a Certificate",
                "operationId": "0ffea9d518b20c405a65d99f6bc3137d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Certificate ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "login",
                "name": "token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}